requirements.txt
requests
pytest
jsonschema
pytest-html
allure-pytest
pytest-rerunfailures